fix: review 코드 누락부분 수정

+++ src/main/java/com/efub/livin/review/service/DormReviewService.java
@@ -8,6 +8,8 @@ import com.efub.livin.review.dto.request.DormReviewCreateRequestDto;
 import com.efub.livin.review.dto.response.DormReviewDetailResponseDto;
 import com.efub.livin.review.dto.response.DormReviewListResponseDto;
 import com.efub.livin.review.repository.DormReviewRepository;
+import com.efub.livin.user.domain.User;
+import com.efub.livin.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,11 +21,16 @@ import java.util.*;
 public class DormReviewService {
 
     private final DormReviewRepository dormReviewRepository;
+    private final UserRepository userRepository;
 
     // 리뷰 생성
     @Transactional
-    public long createDormReview(DormReviewCreateRequestDto request){
-        DormReview review = request.toEntity();
+    public long createDormReview(DormReviewCreateRequestDto request, Long userId){
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+        DormReview review = request.toEntity(user);
 
         //이미지가 있다면 엔티티와 연결
         if(request.getImageUrls() != null) {
@@ -62,10 +69,13 @@ public class DormReviewService {
 
     //리뷰 삭제
     @Transactional
-    public void deleteDormReview(Long id){
+    public void deleteDormReview(Long id, Long loginUserId){
         DormReview review = dormReviewRepository.findById(id)
                 .orElseThrow(()-> new CustomException(ErrorCode.DORM_REVIEW_NOT_FOUND));
 
+        if(!review.getUser().getUserId().equals(loginUserId)){
+            throw new CustomException(ErrorCode.ACCESS_DENIED);
+        }
         dormReviewRepository.delete(review);
 
     }
